Form and Database Update
The form accidentally had incorrect labels for two of the fields. This
has been fixed.

A reference to info.php has been added. This file is only on the live
server and contains $username and $password for the username and
password of our database. We will go over this password again in our
regular meeting this Friday so that everyone in the group can test the
live rendition. Remember to tell me when you push something new so I can
pull it down on the live server.